Core: Deprecate jQuery.parseJSON
Fixes gh-2800
