--- src/ajax/parseJSON.js
@@ -1,9 +1,0 @@
-define( [
-	"../core"
-], function( jQuery ) {
-
-jQuery.parseJSON = JSON.parse;
-
-return jQuery.parseJSON;
-
-} );
