+++ src/deprecated.js
@@ -23,4 +23,6 @@ jQuery.fn.extend( {
 	}
 } );
 
+jQuery.parseJSON = JSON.parse;
+
 } );
